fix: resolve CI failures across lint, SAST, deps, and action versions
- Fix all ruff lint errors (1192 auto-fixed + 26 manual fixes) and
  format 205 files to pass ruff check and ruff format --check
- Add ruff config ignores for ambiguous unicode (RUF001-003), mutable
  ClassVar (RUF012), contextlib.suppress (SIM105), dangling tasks (RUF006)
- Add per-file E402 ignores for mixin/app import patterns
- Fix 5 Bandit high-severity findings:
  - tarfile.extractall → add filter="data" (B202)
  - hashlib.md5 → add usedforsecurity=False (B324 x4)
- Regenerate all requirements lock files with SHA-256 hashes
  (base.txt, dev.txt, metrics.txt) to pass dep-freshness CI job
- Standardize all GitHub Actions to v6 (checkout, setup-python) —
  Dependabot PR #59 only updated some jobs, leaving v4/v5 mix
- Declare asyncio marker in pytest.ini to fix --strict-markers
  compatibility with pytest-asyncio
- Add tests/ to pyproject.toml testpaths

All 209 tests pass. Zero high-severity Bandit findings.

+++ ui/dashboard/architect_critic.py
@@ -58,7 +58,7 @@ def _score_style(score: float) -> str:
 
 
 def _score_bar(score: float, width: int = 20) -> str:
-    filled = int(round(score / 10.0 * width))
+    filled = round(score / 10.0 * width)
     empty = width - filled
     bar = "█" * filled + "░" * empty
     return bar
